osd: erasure_code: Fix re-reads with optimised code path.
When performing optimised reads, prior to this commit, if the experimental
flag was enabled, then recovery would be impossible if it involved re-reading
an existing chunk.

For example. A client performs a read which covers end of chunk A and the start of chunk B. If chunk B fails, then the start of chunk A is required for recovery.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>