include: interval_set: Refactor erase to allow overlaps.
Before this commit, the erase function would assert that intervals to be erased entirely
exist within an existing interval.  This leads to a number of issues:

1. Easy to create hard-to-test edge conditions for client code.
2. Often calling code would do inefficient thing, such a execute intersect, then subtract.
3. It is a particular surprise that this issue also applies to subtract.

This new version of the erase method allows for any sized interval to be used in erase. This
allows for both non or partially overlapping erases, but also can recurse over multiple
contained regions, making it much more convenient to erase.

The new code should be as efficient for existing erases supported by the code.  For a single
non-overlapping erase, the complexity is O(log n) and the constant penalty should be
roughly the same.

For multiple erases, the complexity is somewhere between O(log n) and O(n), scaling with
the number of entries which must be erased.

The complexity of subtract is the same as before O(n log m), but should have a lower constant
than previously held.

I have also removed various workarounds used in interval_set required by the limitations
of the old erase and insert functions.

The added unit tests should be extensive enough to give confidence that the new code has a
functional superset of the old code.

Signed-off-by: Alexander H Ainscow <aainscow@uk.ibm.com>